fix: the self-guardrail blocked its own wave — remediated at the findings
dxkit's dogfood gate blocked PR #134 with four net-new findings; all four
were real and none is cleared by touching the baseline:

- test-file-degradation on spec/requests/smoke_spec.rb: the validation
  battery's FIRST (broken) run leaked a Rails fixture INTO this repo — the
  harness cd'd inside a command substitution, so a case wrote to the repo
  cwd, and a later 'git add -A' swept it into a commit. Removed; the
  hermetic-harness process note already covers the cd-in-subshell trap.
- three large-file findings: the wave grew flow/extract.ts (661),
  sdk/descriptors.ts (578), and model-schema/extract.ts (524) past the
  500-line threshold. Each split along its cohesive seam: the verb-callee +
  resource-expansion branches move to flow/extract-verbs.ts (extract.ts
  keeps the walk and calls in; null-vs-empty return preserves the
  fall-through semantics the client branches rely on), schema-file table
  extraction moves to model-schema/schema-file.ts, and ModelSchemaSupport
  moves to sdk/model-descriptors.ts (barrel re-export — the frozen surface
  is unchanged and the freeze test pins it).

Local guardrail check post-fix: PASSED (0 blocking, the four findings
gone, nothing grandfathered).
