--- spec/requests/smoke_spec.rb
@@ -1,7 +1,0 @@
-describe 'smoke' do
-  it 'gets' do
-    get '/api/articles'
-    post '/api/articles'
-    get '/phantom/route'
-  end
-end
