+++ src/analyzers/model-schema/gather.ts
@@ -34,7 +34,8 @@ import type { ModelSchemaSupport } from '../../languages/types';
 import { withParsedFile } from '../../ast/parse';
 import { grammarShape } from '../../ast/grammar-shape';
 import { walkSourceFiles } from '../tools/walk-source-files';
-import { extractFileModels, extractSchemaFileTables } from './extract';
+import { extractFileModels } from './extract';
+import { extractSchemaFileTables } from './schema-file';
 import { loadSpecModels } from './spec-source';
 import { readSchemaConfig } from './config';
 import { mergePartialEntities } from './model';
